+++ NEWS.md
@@ -11,6 +11,9 @@
 - Give a better error message in `genoprob_to_snpprob()` if `snpinfo`
   is missing the `sdp` column (Issue #207).
 
+- In `read_cross()`, give warnings if there are duplicate column names
+  or duplicate row names in the file.
+
 
 ## qtl2 0.28 (2021-10-11)